fix timing issue with validate publish
