+++ src/c/project-card-edit.js
@@ -31,6 +31,7 @@ const projectCardEdit = {
                         if (!showSuccess()) { showSuccess.toggle(); }
                         if (showError()) { showError.toggle(); }
                         vm.reloadCurrentProject();
+                        railsErrorsVM.validatePublish();
                     }).catch((err) => {
                         if (err.errors_json) {
                             railsErrorsVM.mapRailsErrors(err.errors_json, mapErrors, vm.e);
