+++ src/c/project-goal-edit.js
@@ -1,7 +1,7 @@
 import m from 'mithril';
 import _ from 'underscore';
-import h from '../h';
 import I18n from 'i18n-js';
+import h from '../h';
 import railsErrorsVM from '../vms/rails-errors-vm';
 import projectGoalVM from '../vms/project-goal-vm';
 import popNotification from './pop-notification';
@@ -32,6 +32,7 @@ const projectGoalEdit = {
                     vm.e.resetFieldErrors();
                     if (!showSuccess()) { showSuccess.toggle(); }
                     if (showError()) { showError.toggle(); }
+                    railsErrorsVM.validatePublish();
                 }).catch((err) => {
                     if (err.errors_json) {
                         railsErrorsVM.mapRailsErrors(err.errors_json, mapErrors, vm.e);
